object-particle: Use TVec3<s16>(s16, s16, s16) as in the demo map (clean match)

+++ src/JSystem/JStudio/JStudio_JParticle/object-particle.cpp
@@ -236,9 +236,9 @@ void TAdaptor_particle::TJPACallback_::execute(JPABaseEmitter* emitter) {
         }
         emitter->setGlobalTranslation(pos[0]);
 
-        int angleX = 0x10000 * (pos[1].x / 360.0);
-        int angleY = 0x10000 * (pos[1].y / 360.0);
-        int angleZ = 0x10000 * (pos[1].z / 360.0);
+        s16 angleX = 0x10000 * (pos[1].x / 360.0);
+        s16 angleY = 0x10000 * (pos[1].y / 360.0);
+        s16 angleZ = 0x10000 * (pos[1].z / 360.0);
         emitter->setGlobalRotation(JGeometry::TVec3<s16>(angleX, angleY, angleZ));
         emitter->setGlobalScale(sp0C);
     } else {
